requirements.txt
asgiref==3.2.10
Django==3.0.8
django-rest-multiple-models==2.1.3
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
persiantools==1.5.0
gunicorn==20.0.4
django-cors-headers==3.4.0
